The connections for selected city and saved city were established, and the process of transferring city information based on the selected city was completed.

+++ Havadar/Model/NowWeatherResponse.swift
@@ -16,8 +16,8 @@ struct NowWeatherResponseElement: Codable {
     let ruzgarYon: Int
     let sicaklik: Double
     let veriZamani: String
-    let yagis00Now, yagis10Dk, yagis12Saat, yagis1Saat: Int
-    let yagis24Saat, yagis6Saat: Int
+    let yagis00Now, yagis10Dk, yagis12Saat, yagis1Saat: Double
+    let yagis24Saat, yagis6Saat: Double
     let denizVeriZamani: String
 }
